shadow: R Package for Geometric Shade Calculations

Main functions for calculating:

shadowHeight, Shadow height at individual points or continuous surface

shadowFootprint, Polygonal layer of shadow footprints on the ground

SVF, Sky View Factor (SVF) value at individual points or continuous surface

Typical inputs for these functions include:

location, Queried location(s)

obstacles, A polygonal layer of obstacles (e.g. buildings) outline, with height attributes obstacles_height_field

solar_pos, Solar position (i.e. sun azimuth and elevation angles)

The package also provides functions for related preliminary calculations, such as:

toSeg, Converting polygons to line segments

classifyAz, Finding segment azimuth

shiftAz, Shifting segments by azimuth and distance

ray, Constructing a line between two points